Radius slider moved into its own component file to improve modularity and make for easier testing.
